Create vpsprep.sh
This is a script to automate the process of verifying ports are available, updating the OS, creating the ether1node user, installing ufw, fail2ban, and downloading the ethoFS install script. Users can easily run this script, relog as ether1node user and run the ethoFS install script to complete the process.